rearrange things to work better with pyreverse.
